# api/pkg/routes/openapi-spec.yaml
openapi: 3.0.0
info:
  title: Fampay Backend assignment API
  description: basic search API to search the stored videos using their title and description.
  termsOfService: TODO 
  version:  1.0.0
  contact:
    name: Anuj Parihar
    url: https://github.com/BearTS
    email: anujpflash@gmail.com
servers:
  - url: http://localhost:3000/v1
    description: Local development server
paths:
  /ping:
    get:
      summary: Ping
      operationId: ping
      description: Ping the server
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Pong
  /videos:
    get:
      summary: Get videos
      operationId: getVideos
      description: Get videos
      parameters:
        - name: page
          in: query
          description: page number
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          description: number of videos per page
          required: false
          schema:
            type: integer
        - name: title
          in: query
          description: title of the video
          required: false
          schema:
            type: string
        - name: description
          in: query
          description: description of the video
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetVideosResponse'
                

components:
  schemas:
    Video:
      type: object
      required:
        - id
        - title
        - description
        - publishedAt
        - thumbnail
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        publishedAt:
          type: string
          format: date-time
        thumbnail:
          type: string
    GetVideosResponse:
      type: object
      required:
        - videos
        - page
        - limit
        - total
      properties:
        videos:
          type: array
          items:
            $ref: '#/components/schemas/Video'
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
          format: int64